+++ cli/cli/commands/engine/stop/stop.go
@@ -2,8 +2,8 @@ package stop
 
 import (
 	"context"
-	"github.com/kurtosis-tech/container-engine-lib/lib"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/backend_for_cmd"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
@@ -17,18 +17,22 @@ var StopCmd = &cobra.Command{
 	RunE:  run,
 }
 
+var WithKubernetes bool
+
 func init() {
-	// No flags yet
+	// TODO Remove this in favor of actual Kubernetes info in the config file
+	StopCmd.Flags().BoolVarP(&WithKubernetes, "with-kubernetes", "k", false, "Operate on the engine in kubernetes")
 }
 
 func run(cmd *cobra.Command, args []string) error {
 	ctx := context.Background()
 
+	cmd.Flags().GetBool("with-kubernetes")
 	logrus.Infof("Stopping Kurtosis engine...")
 
-	kurtosisBackend, err := lib.GetLocalDockerKurtosisBackend()
+	kurtosisBackend, err := backend_for_cmd.GetBackendForCmd(WithKubernetes)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend connected to local Docker")
+		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend connected")
 	}
 	engineManager := engine_manager.NewEngineManager(kurtosisBackend)