+++ cli/docs/changelog.md
@@ -1,5 +1,10 @@
 # TBD
 
+# 0.16.1
+### Features
+* Can use kubernetes backend for engine commands through the `--with-kubernetes` flag
+* Upgraded to container-engine-lib which provides new (but still dormant) k8s functionality
+
 # 0.16.0
 ### Changes
 * Upgrade to engine 1.18.0, which uses the enclave data volume (rather than enclave data dirpath) for enclave data storage